+++ src/main/java/duke/command/AddEventCommand.java
@@ -1,10 +1,14 @@
-import java.text.ParseException;
+package duke.command;
 
-public class AddDeadlineCommand extends AddCommand {
+import java.text.ParseException;
+import duke.task.*;
+import duke.io.*;
+
+public class AddEventCommand extends AddCommand {
     private String input;
     private String[] details;
 
-    public AddDeadlineCommand(String input) throws ArrayIndexOutOfBoundsException {
+    public AddEventCommand(String input) throws ArrayIndexOutOfBoundsException {
         super(input);
         details = Parser.getDetails(input);
     }
@@ -12,7 +16,7 @@ public class AddDeadlineCommand extends AddCommand {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws ParseException {
         ui.showLine();
-        taskList.add(new Deadline(details[0], Parser.getAsDate(details[1])));
+        taskList.add(new Event(details[0], Parser.getAsDate(details[1])));
         ui.out("Got it. I've added this task:");
         ui.out(taskList.get(taskList.size() - 1).toString());
         ui.out("Now you have " + taskList.size() + " tasks in the list.");
