+++ src/main/java/duke/command/ListCommand.java
@@ -1,8 +1,13 @@
-public class ExitCommand extends Command {
+package duke.command;
+
+import duke.task.*;
+import duke.io.*;
+
+public class ListCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         ui.showLine();
-        ui.out("Bye. Hope to see you again soon!");
+        ui.list(taskList);
         ui.showLine();
     }
 }
