+++ src/main/java/duke/main/Main.java
@@ -1,3 +1,5 @@
+package duke.main;
+
 public class Main {
     public static void main(String[] args) {
         new Duke("data/duke.txt").run();
