+++ src/main/java/duke/task/Deadline.java
@@ -1,16 +1,19 @@
+package duke.task;
+
 import java.util.Date;
+
 /**
  * Represents a Deadline Task
  */
 public class Deadline extends Task {
     private Date details;
 
-    Deadline(String description, Date details) {
+    public Deadline(String description, Date details) {
         super(description);
         this.details = details;
     }
 
-    Deadline(String description, Date details, boolean done) {
+    public Deadline(String description, Date details, boolean done) {
         super(description, done);
         this.details = details;
     }
@@ -21,7 +24,7 @@ public class Deadline extends Task {
     }
 
     @Override
-    String store(){
+    public String store(){
         return "D|" + getStatus() + "|" + description + "|" + details;
     }
 }
