+++ src/main/java/duke/task/Event.java
@@ -1,16 +1,19 @@
+package duke.task;
+
 import java.util.Date;
+
 /**
  * Represents an Event Task
  */
 public class Event extends Task {
     private Date details;
 
-    Event(String description, Date details) {
+    public Event(String description, Date details) {
         super(description);
         this.details = details;
     }
 
-    Event(String description, Date details, boolean done) {
+    public Event(String description, Date details, boolean done) {
         super(description, done);
         this.details = details;
     }
@@ -21,7 +24,7 @@ public class Event extends Task {
     }
 
     @Override
-    String store(){
+    public String store(){
         return "E|" + getStatus() + "|" + description + "|" + details;
     }
 }
